[Chore #163020832] Refactored UI and Api controller
This allows code base to be scale able and maintain able
for future reviews and adding of new feature to the code
base.

This commit will update exiting code base with new refactored
changes from this chore branch and makes more improvement
to the application user interface and api-endpoints.
Listed below are the files and folder being refactored:
	* Api-endpoints
		- Removed and replaced async with static keyword
		- Moved all controller methods  from object to a single
		class
		- Moved all helper methods to a single class
		- removed all console.log from code base
	* Front-end (UI)
		- Fixed dashboard  min-height issue on larger screens
		- Fixed dashboard sidebar container alignment issues
		- Fixed alignment issues on index page
		  (question-feeds-actions),upvote,downvote,comments
		- Fixed upcoming meetup page issues (attend meetup button)
		  to be click able and link to sign in.

Resolves: https://www.pivotaltracker.com/n/projects/2232110 [#163020832]